split soundcloud
